Ensure that all amount casts are sound (#21)
The currently used type i64 is wide enough for the time being but may not be so in the distant future. Using i128 instead ensures that any u64 value stays represented correctly.

Rosetta represents amounts as strings so no changes to generated code are required.

+++ src/api/account.rs
@@ -41,7 +41,7 @@ impl AccountApi {
                 block_info.block_height.height as i64,
                 block_info.block_hash.to_string(),
             ),
-            vec![amount_from_uccd(account_info.account_amount.microgtu as i64)],
+            vec![amount_from_uccd(account_info.account_amount.microgtu as i128)],
         ))
     }
 }
